CharacterController and ThirdFloor Layout Model Added. New Unity Store Assets Included
I've added the CharacterController and ThirdFloor Layout model to this
update. These changes can be seen in the FloorLayoutDemo. Also, Fowler
sent out the link to our models, so I've included the Office assets and
one of the monsters if you guys want to go ahead and start looking at
them. If there are issues let me know.